launcher {dabtab}	R Documentation

Create a launcher on the desktop for Windows (.bat), Mac (.command), or Linux (.sh)

Description

Create a launcher on the desktop for Windows (.bat), Mac (.command), or Linux (.sh)

Usage


launcher(
  app = c("dabtab", "dabtab.basic"),
  port = 4444,
  pdir = Sys.getenv("HOME")
)

Arguments

app
App to run when the desktop icon is double-clicked ("analytics", "marketing", "quant", or "base"). Default is "analytics"

port
Port to use for shiny::runApp (e.g, port = 4444)

pdir
Project directory to use. Default is Sys.getenv("HOME")

Details

On Windows/Mac/Linux a file named dabtab.bat/dabtab.command/dabtab.sh will be put on the desktop. Double-click the file to launch the specified dabtab app

See Also

win_launcher to create a shortcut on Windows

mac_launcher to create a shortcut on Mac

lin_launcher to create a shortcut on Linux

Examples


## Not run: 
dabtab::launcher("dabtab.model")

## End(Not run)


[Package dabtab version 1.1.0 Index]