mac_launcher {dabtab}	R Documentation

Create a launcher and updater for Mac (.command)

Description

Create a launcher and updater for Mac (.command)

Usage


mac_launcher(
  app = c("dabtab", "dabtab.basic"),
  port = 4444,
  pdir = Sys.getenv("HOME")
)

Arguments

app
App to run when the desktop icon is double-clicked ("dabtab", "dabtab.basic", etc.). Default is "dabtab"

port
Port to use for shiny::runApp (e.g, port = 4444)

pdir
Project directory to use. Default is Sys.getenv("HOME")

Details

On Mac a file named 'dabtab.command' and one named 'update_dabtab.command' will be put on the desktop. Double-click the file to launch the specified dabtab app or update dabtab to the latest version

Examples


## Not run: 
dabtab::mac_launcher()

## End(Not run)


[Package dabtab version 1.1.0 Index]